requirements-test.txt
django-crispy-forms>=1.9.2
coverage
pylint
pylint-django
